+++ mlflow/models/model.py
@@ -3,8 +3,11 @@ import logging
 import os
 import posixpath
 import shutil
+import threading
 import uuid
 import warnings
+from collections import defaultdict
+from contextvars import ContextVar
 from datetime import datetime
 from pathlib import Path
 from pprint import pformat
@@ -1379,3 +1382,33 @@ def set_model(model) -> None:
             pass
 
     raise mlflow.MlflowException(SET_MODEL_ERROR)
+
+
+class _ModelTracker:
+    """
+    Tracks models existing in current context.
+    """
+
+    def __init__(self):
+        self.model_ids: dict[int, str] = {}
+        self._lock = threading.Lock()
+        self._model_locks = defaultdict(threading.Lock)
+
+    def get(self, model: Any) -> Optional[str]:
+        model_id_key = id(model)
+        with self._model_locks[model_id_key]:
+            return self.model_ids.get(model_id_key)
+
+    def set(self, model: Any, model_id: str) -> None:
+        model_id_key = id(model)
+        with self._model_locks[model_id_key]:
+            self.model_ids[model_id_key] = model_id
+
+    def clear(self):
+        with self._lock:
+            self.model_ids.clear()
+            self._model_locks.clear()
+
+
+_MODEL_TRACKER = _ModelTracker()
+_active_model_id = ContextVar("_active_model_id", default=None)
