+++ tests/langchain/conftest.py
@@ -6,7 +6,7 @@ import pytest
 from langchain.embeddings.base import Embeddings
 from pydantic import BaseModel
 
-from mlflow.langchain import _MODEL_TRACKER
+from mlflow.models.model import _MODEL_TRACKER
 
 from tests.helper_functions import start_mock_openai_server
 from tests.tracing.helper import reset_autolog_state  # noqa: F401
